pyproject.toml
[project]
name = "qdrant-mcp-server"
version = "1.0.0"
description = "MCP Server for interacting with Qdrant"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "qdrant-client>=1.7.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "fastmcp>=0.4.0",
]

[project.license]
file = "LICENSE"

[project.scripts]
qdrant-mcp-server = "qdrant_mcp_server:main"

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build" 